# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SEN Migration Dashboard API
  description: |
    REST API for the SEN Migration Platform.
    Provides migration status, object progress, rule analysis,
    reconciliation, test scenario management, forensic extraction,
    process mining, data export, and migration planning.
  version: 1.0.0
  contact:
    name: SEN Team
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:4004
    description: Local development
  - url: https://{host}
    description: Production
    variables:
      host:
        default: sapconnect.example.com
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/testing/generate:
    post:
      tags: [Testing]
      summary: Generate test cases
      operationId: generateTestCases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [description, config, bpmn]
                input:
                  type: string
      responses:
        '200':
          description: Generated test cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  testCases:
                    type: array
                    items:
                      type: object
                  generatedAt:
                    type: string
                    format: date-time
  /api/testing/templates:
    get:
      tags: [Testing]
      summary: List test templates
      operationId: listTestTemplates
      parameters:
        - name: module
          in: query
          required: false
          schema:
            type: string
          description: Filter by SAP module (FI, SD, MM, etc.)
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [smoke, regression, integration]
          description: Filter by test type
      responses:
        '200':
          description: Test templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
  /api/testing/templates/{id}:
    get:
      tags: [Testing]
      summary: Get test template by ID
      operationId: getTestTemplate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Test template detail
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/testing/templates/{id}/instantiate:
    post:
      tags: [Testing]
      summary: Instantiate test template with variables
      operationId: instantiateTestTemplate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                variables:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        '200':
          description: Instantiated test case
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/testing/execute:
    post:
      tags: [Testing]
      summary: Execute test suite
      operationId: executeTestSuite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                testCases:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Test execution results
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: object
                    properties:
                      total:
                        type: integer
                      passed:
                        type: integer
                      failed:
                        type: integer
                  results:
                    type: array
                    items:
                      type: object
                  executedAt:
                    type: string
                    format: date-time
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Required when API_KEY environment variable is set.
        Pass via X-API-Key header or apiKey query parameter.
        Exempt paths: /health, /ready, /metrics
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
...